pyproject.toml
[project]
name = "learninghub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.3",
    "langchain-google-genai>=3.0.1",
    "python-dotenv>=1.2.1",
]
